factor out led stuff into separate unit
Should make later refactoring easier and while at it switch to SPDX
license tags for obvious reasons.
