+++ led.h
@@ -1,0 +1,9 @@
+/*
+ * Copyright (C) 2017 John Crispin <blogic@openwrt.org>
+ *
+ * SPDX-License-Identifier: LGPL-2.1
+ */
+
+#pragma once
+
+void led_add(const char *path, int brightness, int original, int blink, int fade, int on, int off);
